refactor: QueryService, Service 로 분리 - #46
- 책임 분리를 위함
