+++ src/main/java/org/mju_likelion/festival/auth/service/AuthService.java
@@ -7,15 +7,9 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.admin.domain.Admin;
-import org.mju_likelion.festival.admin.domain.repository.AdminJpaRepository;
-import org.mju_likelion.festival.auth.domain.RsaKey;
 import org.mju_likelion.festival.auth.domain.RsaKeyStrategy;
-import org.mju_likelion.festival.auth.dto.request.AdminLoginRequest;
 import org.mju_likelion.festival.auth.dto.request.UserLoginRequest;
-import org.mju_likelion.festival.auth.dto.response.KeyResponse;
 import org.mju_likelion.festival.auth.dto.response.LoginResponse;
-import org.mju_likelion.festival.auth.dto.response.TermResponse;
 import org.mju_likelion.festival.auth.util.jwt.AuthenticationRole;
 import org.mju_likelion.festival.auth.util.jwt.JwtUtil;
 import org.mju_likelion.festival.auth.util.jwt.Payload;
@@ -40,24 +34,11 @@ public class AuthService {
   private final MjuApiUtil mjuApiUtil;
   private final UserJpaRepository userJpaRepository;
   private final TermJpaRepository termJpaRepository;
-  private final AdminJpaRepository adminJpaRepository;
   private final JwtUtil jwtUtil;
-
-  public KeyResponse getKey() {
-    RsaKeyManager rsaKeyManager = rsaKeyManager();
-
-    RsaKey rsaKey = rsaKeyManager.generateRsaKey();
-
-    String credentialKey = rsaKeyManager.savePrivateKey(rsaKey.privateKey());
-
-    RsaKeyStrategy rsaKeyStrategy = rsaKeyManager.rsaKeyStrategy();
-
-    return new KeyResponse(rsaKey.publicKey(), credentialKey, rsaKeyStrategy);
-  }
 
   @Transactional
   public LoginResponse userLogin(UserLoginRequest userLoginRequest, RsaKeyStrategy rsaKeyStrategy) {
-    RsaKeyManager rsaKeyManager = rsaKeyManager(rsaKeyStrategy);
+    RsaKeyManager rsaKeyManager = rsaKeyManagerContext.rsaKeyManager(rsaKeyStrategy);
 
     String key = userLoginRequest.getKey();
     String studentId = rsaKeyManager.decryptByKey(userLoginRequest.getEncryptedStudentId(), key);
@@ -69,41 +50,6 @@ public class AuthService {
 
     String accessToken = jwtUtil.create(new Payload(userId, AuthenticationRole.USER));
     return new LoginResponse(accessToken);
-  }
-
-  @Transactional(readOnly = true)
-  public LoginResponse adminLogin(AdminLoginRequest adminLoginRequest,
-      RsaKeyStrategy rsaKeyStrategy) {
-    RsaKeyManager rsaKeyManager = rsaKeyManager(rsaKeyStrategy);
-
-    String key = adminLoginRequest.getKey();
-    String loginId = rsaKeyManager.decryptByKey(adminLoginRequest.getEncryptedLoginId(), key);
-    String password = rsaKeyManager.decryptByKey(adminLoginRequest.getEncryptedPassword(), key);
-
-    Admin admin = getExistingAdmin(loginId, password);
-
-    String accessToken = jwtUtil.create(
-        new Payload(admin.getId(), AuthenticationRole.from(admin.getRole())));
-    return new LoginResponse(accessToken);
-  }
-
-  @Transactional(readOnly = true)
-  public List<TermResponse> getTerms() {
-    return termJpaRepository.findTermsByOrderBySequenceAsc().stream()
-        .map(TermResponse::of)
-        .collect(Collectors.toList());
-  }
-
-  /**
-   * 관리자가 존재하는지 확인하고, 존재하는 경우 사용자를 반환한다.
-   *
-   * @param loginId  로그인 ID
-   * @param password 비밀번호
-   * @return 관리자
-   */
-  private Admin getExistingAdmin(String loginId, String password) {
-    return adminJpaRepository.findByLoginIdAndPassword(loginId, password)
-        .orElseThrow(() -> new UnauthorizedException(ErrorType.INVALID_CREDENTIALS_ERROR));
   }
 
   /**
@@ -178,13 +124,5 @@ public class AuthService {
     if (!mjuApiUtil.doUserCheck(studentId, password)) {
       throw new UnauthorizedException(ErrorType.INVALID_CREDENTIALS_ERROR);
     }
-  }
-
-  private RsaKeyManager rsaKeyManager() {
-    return rsaKeyManagerContext.rsaKeyManager();
-  }
-
-  private RsaKeyManager rsaKeyManager(RsaKeyStrategy rsaKeyStrategy) {
-    return rsaKeyManagerContext.rsaKeyManager(rsaKeyStrategy);
   }
 }
